+++ index/fenwicktree/BUILD.bazel
@@ -1,8 +1,10 @@
 load("@rules_rust//rust:defs.bzl", "rust_doc_test", "rust_library", "rust_test", "rust_test_suite")
 
 rust_library(
-    name = "fenwicktree",
-    srcs = ["src/lib.rs"],
+    name = "fenwicktree_rs",
+    srcs = ["fenwicktree.rs"],
+    crate_name = "fenwicktree",
+    crate_root = "fenwicktree.rs",
     visibility = ["//visibility:public"],
     deps = [
         "//bits:bits_rs",
@@ -10,18 +12,18 @@ rust_library(
 )
 
 rust_test(
-    name = "cfg_test",
-    crate = ":fenwicktree",
+    name = "fenwicktree_rs_lib_test",
+    crate = ":fenwicktree_rs",
 )
 
 rust_doc_test(
-    name = "doc_test",
-    crate = ":fenwicktree",
+    name = "fenwicktree_rs_doc_test",
+    crate = ":fenwicktree_rs",
 )
 
 rust_test_suite(
-    name = "integration",
-    srcs = glob(["tests/**"]),
+    name = "fenwicktree_rs_tests",
+    srcs = glob(["*_test.rs"]),
     proc_macro_deps = [
         "@crates//:quickcheck_macros",
     ],
@@ -29,7 +31,7 @@ rust_test_suite(
         "-Coverflow-checks=off",
     ],
     deps = [
-        ":fenwicktree",
+        ":fenwicktree_rs",
         "//bits:bits_rs",
         "@crates//:quickcheck",
     ],
